chore: move class to package
