+++ marketplace-data/src/main/kotlin/dev/ja/marketplace/data/trackers/PaymentAmountTracker.kt
@@ -1,9 +1,9 @@
 /*
- * Copyright (c) 2023 Joachim Ansorg.
+ * Copyright (c) 2023-2024 Joachim Ansorg.
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-package dev.ja.marketplace.data
+package dev.ja.marketplace.data.trackers
 
 import dev.ja.marketplace.client.Amount
 import dev.ja.marketplace.client.Marketplace
@@ -11,6 +11,9 @@ import dev.ja.marketplace.client.YearMonthDay
 import dev.ja.marketplace.client.YearMonthDayRange
 import java.math.BigDecimal
 
+/**
+ * Tracks the payment within a given range of dates.
+ */
 data class PaymentAmountTracker(val filterDateRange: YearMonthDayRange) {
     private var _totalAmountUSD: Amount = BigDecimal.ZERO
     private var _feesUSD: Amount = BigDecimal.ZERO
